add conditional to not drop db when fails for testing purposes

+++ scripts/browser_tests.sh
@@ -81,8 +81,11 @@ docker-compose -f ${DOCKER_COMPOSE_FILE} run --rm \
 EXIT_CODE=$?
 
 # TEST CLEANUP
-docker exec -t -u postgres productcatalog_db_1 \
-  psql --command "DROP DATABASE IF EXISTS test_${DB_NAME};"
+if [ ${EXIT_CODE} != 1 ]
+then
+  docker exec -t -u postgres productcatalog_db_1 \
+    psql --command "DROP DATABASE IF EXISTS test_${DB_NAME};"
+fi
 
 docker-compose -f ${DOCKER_COMPOSE_FILE} stop
